+++ infrastructure/src/main/java/com/music/s3/CustomObjectMetadataProvider.java
@@ -3,8 +3,8 @@ package com.music.s3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.transfer.ObjectMetadataProvider;
 import com.music.constatns.AudioQuality;
-import com.music.upload.dto.MusicUpload;
-import com.music.upload.dto.LyricsUpload;
+import com.music.dto.upload.MusicUpload;
+import com.music.dto.upload.LyricsUpload;
 import java.io.File;
 import java.util.Iterator;
 import java.util.Map;
