+++ music/src/main/java/com/music/controller/MusicController.java
@@ -1,11 +1,10 @@
 package com.music.controller;
 
-import com.music.upload.dto.MusicUploadResponse;
-import com.music.upload.dto.MusicUploadRequest;
-import com.music.upload.service.MusicUploadService;
+import com.music.dto.upload.MusicUploadResponse;
+import com.music.dto.upload.MusicUploadRequest;
+import com.music.service.upload.MusicUploadService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
