+++ music/src/main/java/com/music/convert/AudioQualityConverter.java
@@ -1,4 +1,4 @@
-package com.music.upload.convert;
+package com.music.convert;
 
 import com.music.adaptor.AudioConverter;
 import com.music.constatns.AudioQuality;
